refactor(shell): extract CanvasInteractionSurface component
